+++ src/test/resources/pta/taint/SimpleTaint.java
@@ -14,5 +14,14 @@ class SimpleTaint {
 
         String s4 = Sanitizer.sanitize(s1);
         SourceSink.sink(s4); // no taint
+
+        String s5 = SourceSink.tainted1;
+        SourceSink.sink(s5); // taint
+
+        String s6 = SourceSink.untainted;
+        SourceSink.sink(s6); // no taint
+
+        String s7 = new SourceSink().tainted2;
+        SourceSink.sink(s7); // taint
     }
 }